Changed use CarpetRoll rather the four separate classes for holding the data.
Worked on MainForm

+++ Mahlo2/Models/IMahloRoll.cs
@@ -12,7 +12,7 @@
     int RollLength { get; }
     double RollWidth { get; }
     string DefaultRecipe { get; }
-    decimal PatternRepeatLength { get; }
+    double PatternRepeatLength { get; }
     string ProductImageURL { get; }
 
     int Feet { get; set; }
